Create Listing 4.11  The intconv.c Program
It's very important to keep in mind that mismatches can cause problems. It's a little bit complex to explain how does this program works when you mismatch the conversion specification to the type, you can check it carefully on the book, or, if you need to, you can ask me about it.