Create Listing 4.14  The longstrg.c Program
Method 1 is to use more than one printf() statement.
Method 2 is to terminate the end of the first line with a backslash/return combination. The effect is to continue the string over to the next line. Remember, the next line has to start at the far left, as shown.
Method 3, which ASCI C introduced, is string concatenation. If you follow one quoted string constant with another, separated only by whitespace, C treats the combination as a single string.