Create Listing 4.15  The input.c Program
The scanf() function uses whitespace (newlines, tabs, and spaces) to decide how to divide the input into separate fields. It matches up consecutive conversion specifications to consecutive fields, skipping over the whitespace in between. (The only exception to this is the %c specification, which reads the very next characters, even if that character is whitespace.)