Create Listing  5.8  The sizeof.c Program
The sizeof operator returns the size, in bytes, of its operand.
The C header files system can use typedef to make size_t a synonym for unsigned int on one system or for unsigned long on another. Thus, when you use the size_t type, the compiler will substitute the standard type that works for your system.