Create Listing  5.9  The min_sec.c Program
This program uses scanf() to fetch new values for the variable sec. As long as the value is positive, the loop continues. When the user enters a zero or negative value, the loop quits. The important point is that each loop cycle revises the value of the variable being tested.