Create Listing 5.15  The pound.c Program
Variable names are private to the function. This means that a name defined in one function doesn't conflict with the same name defined elsewhere.
The function call "pound(times)" causes the "times" values of 5 to be assigned to n.
The prototype informs the compiler that pound() expects an int argument. In response, when the compiler reaches the pound(ch) expression, it automatically applies a typecast to the ch argument, converting it to an int argument.
Similarly, the last call, pound(f), generates a type cast to convert the type float variable f to the proper type for this argument.