Create Listing 6.8  The trouble.c Program
The value of an assignment statement is the value of the left side, so "status = 1" has the same numerical value of 1, which makes the loop never quits.
When scanf() fails to read the specified form of input, it leaves the nonconforming input in place to be read the next time. When scanf() tries to read the q as an integer and fails, it leaves the q there. During the next loop cycle, scanf() attempts to read where it left off the lase time -- at the q. Once again, scanf() fails to read the q as an integer, so not only does this example set up an infinite loop, it also creates a loop of infinite failure.