Rename All codes/Chapter 6  C Control Statements -- Looping/Listing 6.13  The postage.c Program to All codes/Chapter  6  C Control Statements -- Looping/Listing 6.13  The postage.c Program
The comma operator extends the flexibility of the loop by enabling you to include more than one initialization or update expression in a single for loop specification.
It also has two further properties. First, it guarantees that the expression it separates are evaluated in a left-to-right order. Second, the value of the whole comma expression is the value of the right-hand member.