Create Listing  7.2  The cypher1.c Program
The getchar() function takes no arguments, and it returns the next character from input; the putchar() function prints its argument. Because these functions deal only with characters , they are faster and more compact than the more general scanf() and printf() functions, and they don't need format specifiers.
This program uses the value of the input item itself to decide when to terminate the loop.