Create Listing  7.4  The electric.c Program
This program uses symbolic constants for the rates and the rate breakpoints so that the constants are conveniently gathered in one place. If the power company changes its rates or rate breakpoints, having the rates and rate breakpoints in one place makes them easy to update.
BASE1 and BASE2 are expressed in terms of the rates and breakpoints. Then, if the rates or breakpoints change, the bases are updated automatically.
The program consists of an if else statement for which the statement part of the else is another if else statement.