Create Listing  7.8  The paint.c Program
You can use the conditional expression "?:" when you have a variable to which you want to assign one of two possible values. Usually, an "if else" statement can accomplish the same thing as the conditional operator. The conditional operator version, however, is more compact and, depending on the compiler, may result in more compact program code.
The conditional operator can use strings for its second and third operands.