Create Listing 8.4  The guess.c Program
The program makes two guesses every time you enter 'n', the reason is that the program reads the 'n' response as a denial that the number is 1 and then reads the newline character as a denial that the number is 2.
One solution is to use a while loop to discard the rest of the input line, including the newline character. (This has the additional merit of treating responses such as 'no' and 'no way' the same as a simple 'n'.