Create Listing 9.12  The loccheck.c Program
We can see from the output that the four have different addresses, and the computer considers them to be four separate variables. Also, the call mikado(pooh) did convey the value (2) of the actual argument (pooh of main()) to the formal argument (bah of mikado()). Note that just the value was transferred, the two variables involved (pooh of main() and bah of mikado()) retain their distinct identities.