Create Listing  10.4  The day_mon2.c Program
When you use empty brackets to initialize an array, the compiler counts the number of items in the list and makes the array that large.
The sizeof operator gives the size, in bytes, of the object, or type, following it. Dividing the size of the entire array by the size of one element tells us how many elements are in the array.