Rename Listing 5.14  The convert.c Program to Listing 5.14  The convert.c Program.c
The character 'C' is stored as a 1-byte ASCII value in ch, the integer variable i receives the integer conversion of 'C', which is 67 stored as 4 bytes, and fl receives the floating conversion of 67, which is 67.00.
This program tries a case of demotion, setting ch equal to an out-of-range number. After the extra bits are ignored, ch winds up with the ASCII code for the S character. (1107 % 256 == 83, the code for S)
This program tries another case of demotion, setting ch equal to a floating point number. After truncation takes place, ch winds up with the ASCII code for the P character.