Rename Listing  7.7  The wordcnt.c Program to Listing  7.7  The wordcnt.c Program.c
"If c is not whitespace and if you are not in a word." If both conditions are met, you must be starting a new word, and n_words is incremented. If you are in the middle of a word, the first condition holds, but "inword" will be true, and n_words is not incremented. When you reach the next whitespace character, "inword" is set equal to false again.