Rename Listing 4.16  The varwid.c Program.c to Listing 4.16  varwid.c
Suppose that you don't want to commit yourself to a field width in advance but rather you want the program to specify it. You can do this by using * instead of a number for the field width, but you also have to add an argument to tell what the field should be.