Rename Listing  7.5  The divisors.c Program.c to Listing  7.5  divisors.c
Traditionally, C has used the int type for flags, but the new _Bool type matches the requirements perfectly. Furthermore, by including the stdbool.h header file, you can use bool instead of the keyword _Bool for the type and use the identifiers true and false instead of 1 and 0.
The program uses the comma operator in the for loop control expression to enable you to initialize isPrime to true for each new input number.