Rename Listing  7.6  The chcount.c Program.c to Listing  7.6  chcount.c
You can translate the if statement as, "If the character is not a double quote AND if it is not a single quote, increase charcount by 1." Both conditions must be true if the whole expression is to be true. The logical operators have a lower precedence than the relational operators, so it is not necessary to use additional parentheses to group the subexpressions.