Rename Listing  7.9  The skippart.c Program.c to Listing  7.9  skippart.c
The continue statement causes the program to skip over the rest of the loop, which is devoted to processing valid input. Instead, the program starts the next loop cycle by attempting to read the next input value.
One way to avoid using continue is omitting the continue and making the remaining part of the loop an else block, but an advantage of using continue in this case is that you can eliminate one level of indentation in the main group of statements. Being concise can enhance readability when the statements are long or are deeply nested already.