Rename Listing 8.2  The echo_eof.c Program.c to Listing 8.2  echo_eof.c
You don't have to define EOF because stdio.h takes care of that.
On most Unix and Linux systems, for example, pressing Ctrl+D at the beginning of a line causes the end-of-file signal to be transmitted. Many micro-computing systems recognize Ctrl+Z at the beginning of a line as an end-of-file signal; some interpret a Ctrl+Z anywhere as an end-of-file signal.