Rename Listing 8.5  The showchar1.c Program.c to Listing 8.5  showchar1.c
The getchar() doesn't skip over newline characters, so the newline character at the end of the first input line is read by getchar() during the next cycle of the loop before you have a chance to enter anything else. Then it's assigned to 'ch', and 'ch' being the newline character is the condition that terminates the loop.