Create Listing 10.11  sum_arr2.c
Because the "while (start < end)" test is for inequality, the last element processed is the one just before the element pointed by "end". This means that "end" actually points to the location after the final element in the array. (C guarantees that when it allocates space for an array, a pointer to the first location after the end of the array is a valid pointer, and that makes constructions such as this one valid.)