Create Listing  11.2  strptr.c
The %p format produces an address. So if the phrase "are" is an address, then %p should print the address of the first character in the string.
The *"space farers" should produce the value to which the address points, which should be the first character of the string "space farers".