Create Listing  11.7  fgets1.c
The fgets() function is designed for file input. It takes a second argument indicating the maximum number of characters to read. If this argument has the value n, fgets() reads up to n-1 characters or through the newline character, whichever comes first. If fgets() reads the newline, it stores it in the string, unlike gets(), which discards it. It takes a third argument indicating which file to read. To read from the keyboard, use stdin (for standard input) as the argument; this identifier is defined in stdio.h.

Because the fgets() function includes the newline as part of the string (assuming the input line fits, it's often paired with fputs(), which works like puts(), except that it doesn't automatically append a newline. It takes a second argument to indicate which file to write to. For the computer monitor we can use stdout (for standard output as an argument.