Create Listing  11.8  fgets2.c
The system uses buffered I/O. This means the input is stored in temporary memory (the buffer until the Return key is pressed; this adds a newline character to the input and sends the whole line on to fgets(). On output, fputs() sends characters to another buffer, and when a newline is sent, the buffer contents are sent on to the display.