Create Listing 11.17  test_fit.c
The fit() function placed a '\0' character in the 39th element of the array, replacing a comma. The puts() function stops at the first null character and ignores the rest of the array.
The expression mesg +39 is the address of mesg[39], which is a space character. So puts() displays that character and keeps going until it runs into the original null character.