Create Listing 11.22  compback.c
strcmp() returns a negative number if the first string precedes the second alphabetically and that it returns a positive number if the order is the other way.

If the initial characters are identical, strcmp() moves along until it finds the first pair of disagreeing characters. It then returns the corresponding code.