Create Listing 12.14  dyn_arr.c
The exit() function, prototyped in stdlib.h, is called to terminate the program if memory allocation fails. The standard provides for two return values that are guaranteed to work with all operating systems: EXIT_SUCCESS to indicate normal program termination, and EXIT_FAILURE to indicate abnormal termination.

The free() function frees only the block of memory to which its argument points. Some operating systems will free allocated memory automatically when a program finishes, but others may not. So use free() and don't rely on the operating system to clean up for you.