Create Listing  13.1  count.c
The program checks the value of argc to see if there is a command-line argument. If there isn't, the program prints a usage message and exits.

Different exit values can be used to distinguish between different causes of failure.

exit() will terminate the program immediately in any situation, while return passes control to the previous level.