Create Listing  13.2  reducto.c
The 'LEN - 5' argument leaves room for the '.red' suffix and the final null character. No null character is copied if the argv[1] string is longer than 'LEN - 5', so the program adds a null character just in case. The first null character in 'name' after the strncpy() call then is overwritten by the period in '.red' when the strcat() function appends that string.