Create Listing  13.4  reverse.c
The first of the three arguments to fseek() is a FILE pointer to the file being searched.

The second argument to fseek() is called the offset. This argument tells how far to move from the starting point.

The third argument is the mode, and it identifies the starting point.

The ftell() function is type long, and it returns the current file location.