Create Listing  14.4  friends.c
Unlike the case for arrays, the name of a structure is not the address of the structure; you need to use the & operator.

A structure pointer followed by the -> operator works the same way as a structure name followed by the . (dot) operator. (You can't say 'him.income' because 'him' is not a structure name. Note that 'him' is a pointer, 'him->income' is a member of the pointed-to structure.)

In summary, if 'him' is a pointer to a type 'guy' structure named 'barney', the following are all equivalent:
barney.income == (*him).income == him->income      // assuming him == &barney