Create Listing  14.7  funds3.c
When 'sum()' is called, an automatic variable called 'moolah' is created according to the 'funds' template. The members of this structure are then initialized to be copies of the values held in the corresponding members of the structure 'stan'. Therefore, the computations are done by using a copy of the original structure; whereas, the precedence program (the one using a pointer) used the original structure.

Because 'moolah' is a structure, the program uses 'moolah.bankfund', not 'moolah->bankfund'. On the other hand, Listing 14.6 used 'money->bankfund' because 'money' is a pointer, not a structure.