Create Listing 14.14  booksave.c
The "a+b" mode is used for opening the file. The a+ part lets the program read the whole file and append data to the end of the file. The b is the ANSI way of signifying that the program will use the binary file format.

We chose the binary mode because fread() and fwrite() are intended for binary files. True, some of the structure contents are text, but the 'value' member is not. If you use a text editor to look at 'book.dat', the text part will show up okay, but the numeric part will be unreadable and could even cause your text editor to barf.

The rewind() command ensures that the file position pointer is situated at the start of the file, ready for the first read.