Create Listing 14.15  enum.c
The purpose of enumerated types is to enhance a program's readability and make it easier to maintain.

Note that the enumerated types are for internal use.

Because the enumerated type is an integer type, enum variables can be used in expressions in the same manner as integer variables. They make convenient labels for a case statement.