Create Listing  15.1  ninbit.c
The CHAR_BIT macro from limits.h represents the number of bits in char. The sizeof operator returns the size in terms of char, so the expression CHAR_BIT * sizeof(int) is the number of bits in an int. The 'bin_str' array has that many elements plus 1 to allow for the terminating null character.

The 'show_bstr()' function breaks up the bits into groups of four to make the string easier to read.