Create Listing  15.5  align.c
Alignment refers to how objects are positioned in memory. The _Alignof operator yields the alignment requirement of a type.

Alignment values should be a non-negative integer power of two. Bigger alignment values are termed stricter or stronger than smaller ones, while smaller ones are termed weaker.

You can use the _Alignof specifier to request a specific alignment for a variable or type. But you shouldn't request an alignment weaker than the fundamental alignment for the type.