Create Listing  16.9  ifdef.c
The #ifdef directive says that if the following identifier has been defined by the preprocessor, follow all the directives and compile all the C code up to the next #else or #endif, whichever comes first. If there is an #else, everything from the #else to the #endif is done if the identifier isn't defined.