Create Listing 16.16  byebye.c
atexit() function uses function pointers! To use the atexit() function, simply pass it the address of the function you want called on exit (The name of a function acts as an address when used as a function argument). Then atexit() registers that function in a list of functions to be executed when exit() is called. When the exit() function is finally called, it executes these functions, with the last function added being executed first. Typically, the functions registered by atexit() would perform housekeeping tasks, such as updating a program-monitoring file or resetting environmental variables.

After exit() executes the functions specified by atexit(), it does some tidying of its own. It flushes all output streams, closes all open streams, and closes temporary files created by calls to the standard I/O function tmpfile(). Then exit() returns control to the host environment and, if possible, reports a termination status to the environment.